Moved the Spinner to the shared folder
